Remove state trie migration to kusama, add it to polkadot
