+++ relay/kusama/Cargo.toml
@@ -84,7 +84,6 @@ pallet-session = { default-features = false , version = "25.0.0" }
 pallet-society = { default-features = false, version = "25.0.0" }
 frame-support = { default-features = false , version = "25.0.0" }
 pallet-staking = { default-features = false , version = "25.0.0" }
-pallet-state-trie-migration = { default-features = false , version = "26.0.0" }
 pallet-staking-runtime-api = { default-features = false , version = "11.0.0" }
 frame-system = { default-features = false , version = "25.0.0" }
 frame-system-rpc-runtime-api = { default-features = false , version = "23.0.0" }
